requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-core
redis
psycopg2-binary
elasticsearch
python-dotenv
pytest
pytest-cov
black
isort
flake8
mypy
pre-commit
pyyaml
python-jose[cryptography]
pydantic-settings
aioredis
httpx
